requirements.txt
Django==1.9.12

django-geoposition==0.3.0

django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11


ipdb==0.11

ipython==6.2.1
